chore: enhance fetch all products
