+++ actions/lib/aem.js
@@ -10,7 +10,7 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-const { request } = require('../../utils');
+const { request } = require('../utils');
 
 /**
  * Creates an instance of AdminAPI.
